refactor(logger-config): split into multiple services
