+++ src/features/logger/enums/logger-config-value-name.enum.spec.ts
@@ -1,0 +1,15 @@
+import { LoggerConfigValueNameEnum } from "./logger-config-value-name.enum";
+
+describe(`LoggerConfigValueNameEnum`, (): void => {
+  it(`should have a member "IS_ENABLED"`, (): void => {
+    expect.assertions(1);
+
+    expect(LoggerConfigValueNameEnum.IS_ENABLED).toStrictEqual(`enable state`);
+  });
+
+  it(`should have a member "LEVEL"`, (): void => {
+    expect.assertions(1);
+
+    expect(LoggerConfigValueNameEnum.LEVEL).toStrictEqual(`level`);
+  });
+});
