+++ src/features/logger/enums/logger-config-value-name.enum.ts
@@ -1,0 +1,4 @@
+export enum LoggerConfigValueNameEnum {
+  IS_ENABLED = `enable state`,
+  LEVEL = `level`,
+}
